refactor: split useVirtualScroll into 3 modules (T10)
useVirtualScroll (743 → 354 lines orchestrator) decomposed into:
- virtualScrollHeightCache.js: pure height cache with prefix-sum
  (createHeightCache factory, ~137 lines) — zero Vue reactivity deps
- useVirtualScrollNavigation.js: scroll-to operations composable
  (scrollToAnchor, scrollToBottom, scrollToIndex, getScrollAnchor,
   programmatic scroll sequencing, ~274 lines)
- useVirtualScroll.js: orchestrator with refs, observers, scroll
  handler, watchers, lifecycle (~354 lines)

Also includes bug fixes from audit:
- Sequence counter for isProgrammaticScrolling (no concurrent corruption)
- All setTimeout calls tracked and cleared on unmount
- mounted guard prevents post-unmount DOM mutations
- Prune stale heights on item changes
- Prefix-sum O(1) spacer lookups
- Viewport-aware renderStart instead of hardcoded 20
- Items count validation in scrollToAnchor/scrollToIndex
